Upgrade minimum supported version to 21.8 in ClickHouse
Staying in version 21.8 (EOL is 31 Aug 2022) because
community members may still use the version.
