+++ docs/src/main/sphinx/connector/clickhouse.rst
@@ -16,7 +16,7 @@ Requirements
 
 To connect to a ClickHouse server, you need:
 
-* ClickHouse (version 21.3 or higher) or Altinity (version 20.8 or higher).
+* ClickHouse (version 21.8 or higher) or Altinity (version 20.8 or higher).
 * Network access from the Trino coordinator and workers to the ClickHouse
   server. Port 8123 is the default port.
