+++ plugin/trino-clickhouse/src/test/java/io/trino/plugin/clickhouse/TestClickHouseLatestConnectorTest.java
@@ -16,8 +16,6 @@ package io.trino.plugin.clickhouse;
 import com.google.common.collect.ImmutableMap;
 import io.trino.testing.QueryRunner;
 
-import java.util.OptionalInt;
-
 import static io.trino.plugin.clickhouse.ClickHouseQueryRunner.createClickHouseQueryRunner;
 import static io.trino.plugin.clickhouse.TestingClickHouseServer.CLICKHOUSE_LATEST_IMAGE;
 
@@ -36,33 +34,5 @@ public class TestClickHouseLatestConnectorTest
                         .put("clickhouse.map-string-as-varchar", "true")
                         .buildOrThrow(),
                 REQUIRED_TPCH_TABLES);
-    }
-
-    @Override
-    protected OptionalInt maxTableNameLength()
-    {
-        // The numeric value depends on file system
-        return OptionalInt.of(255 - ".sql.detached".length());
-    }
-
-    @Override
-    protected String errorMessageForCreateTableAsSelectNegativeDate(String date)
-    {
-        // Override because the DateTime range was expanded in version 21.4 and later
-        return "Date must be between 1970-01-01 and 2149-06-06 in ClickHouse: " + date;
-    }
-
-    @Override
-    protected String errorMessageForInsertNegativeDate(String date)
-    {
-        // Override because the DateTime range was expanded in version 21.4 and later
-        return "Date must be between 1970-01-01 and 2149-06-06 in ClickHouse: " + date;
-    }
-
-    @Override
-    protected String errorMessageForDateYearOfEraPredicate(String date)
-    {
-        // Override because the DateTime range was expanded in version 21.4 and later
-        return "Date must be between 1970-01-01 and 2149-06-06 in ClickHouse: " + date;
     }
 }
